refactor: 커스텀 체크리스트 질문 조회 메서드 이동
