+++ backend/bang-ggood/src/main/java/com/bang_ggood/checklist/controller/ChecklistController.java
@@ -8,7 +8,7 @@ import com.bang_ggood.checklist.service.ChecklistManageService;
 import com.bang_ggood.checklist.service.ChecklistService;
 import com.bang_ggood.question.dto.request.CustomChecklistUpdateRequest;
 import com.bang_ggood.question.dto.response.CategoryCustomChecklistQuestionsResponse;
-import com.bang_ggood.question.dto.response.ChecklistQuestionsResponse;
+import com.bang_ggood.question.dto.response.CustomChecklistQuestionsResponse;
 import com.bang_ggood.user.domain.User;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
@@ -45,9 +45,10 @@ public class ChecklistController {
         return ResponseEntity.noContent().build();
     }
 
+    // TODO : 엔드포인트 통일 with CustomChecklist
     @GetMapping("/checklists/questions")
-    public ResponseEntity<ChecklistQuestionsResponse> readChecklistQuestions(@AuthPrincipal User user) {
-        return ResponseEntity.ok(checklistService.readChecklistQuestions(user));
+    public ResponseEntity<CustomChecklistQuestionsResponse> readCustomChecklistQuestions(@AuthPrincipal User user) {
+        return ResponseEntity.ok(checklistManageService.readCustomChecklistQuestions(user));
     }
 
     @GetMapping("/checklists/{id}")
