+++ backend/bang-ggood/src/main/java/com/bang_ggood/question/dto/response/CustomChecklistQuestionsResponse.java
@@ -1,0 +1,6 @@
+package com.bang_ggood.question.dto.response;
+
+import java.util.List;
+
+public record CustomChecklistQuestionsResponse(List<CategoryQuestionsResponse> categories) {
+}
