+++ backend/bang-ggood/src/test/java/com/bang_ggood/auth/service/AuthServiceTest.java
@@ -3,12 +3,13 @@ package com.bang_ggood.auth.service;
 import com.bang_ggood.IntegrationTestSupport;
 import com.bang_ggood.auth.dto.request.OauthLoginRequest;
 import com.bang_ggood.checklist.dto.response.UserChecklistsPreviewResponse;
+import com.bang_ggood.checklist.service.ChecklistManageService;
 import com.bang_ggood.checklist.service.ChecklistService;
 import com.bang_ggood.global.exception.BangggoodException;
 import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.question.domain.Question;
 import com.bang_ggood.question.dto.response.CategoryQuestionsResponse;
-import com.bang_ggood.question.dto.response.ChecklistQuestionsResponse;
+import com.bang_ggood.question.dto.response.CustomChecklistQuestionsResponse;
 import com.bang_ggood.user.UserFixture;
 import com.bang_ggood.user.domain.User;
 import com.bang_ggood.user.repository.UserRepository;
@@ -37,6 +38,8 @@ class AuthServiceTest extends IntegrationTestSupport {
     private AuthService authService;
     @Autowired
     private ChecklistService checklistService;
+    @Autowired
+    private ChecklistManageService checklistManageService;
     @Autowired
     private UserRepository userRepository;
     @Autowired
@@ -83,10 +86,10 @@ class AuthServiceTest extends IntegrationTestSupport {
 
         // then
         User user = authService.extractUser(token);
-        ChecklistQuestionsResponse checklistQuestions = checklistService.readChecklistQuestions(user);
+        CustomChecklistQuestionsResponse customChecklistQuestions = checklistManageService.readCustomChecklistQuestions(user);
 
         int sum = 0;
-        for (CategoryQuestionsResponse response : checklistQuestions.categories()) {
+        for (CategoryQuestionsResponse response : customChecklistQuestions.categories()) {
             sum += response.questions().size();
         }
