+++ backend/bang-ggood/src/test/java/com/bang_ggood/checklist/service/ChecklistServiceTest.java
@@ -11,13 +11,6 @@ import com.bang_ggood.global.exception.ExceptionCode;
 import com.bang_ggood.like.domain.ChecklistLike;
 import com.bang_ggood.like.repository.ChecklistLikeRepository;
 import com.bang_ggood.option.repository.ChecklistOptionRepository;
-import com.bang_ggood.question.CustomChecklistFixture;
-import com.bang_ggood.question.domain.CustomChecklistQuestion;
-import com.bang_ggood.question.domain.Question;
-import com.bang_ggood.question.dto.response.CategoryQuestionsResponse;
-import com.bang_ggood.question.dto.response.ChecklistQuestionsResponse;
-import com.bang_ggood.question.dto.response.QuestionResponse;
-import com.bang_ggood.question.repository.CustomChecklistQuestionRepository;
 import com.bang_ggood.room.RoomFixture;
 import com.bang_ggood.room.domain.Structure;
 import com.bang_ggood.room.repository.RoomRepository;
@@ -27,10 +20,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import java.util.Collection;
 import java.util.List;
 
-import static com.bang_ggood.question.CustomChecklistFixture.CUSTOM_CHECKLIST_QUESTION_DEFAULT;
 import static com.bang_ggood.user.UserFixture.USER1;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
@@ -50,9 +41,6 @@ class ChecklistServiceTest extends IntegrationTestSupport {
 
     @Autowired
     private RoomRepository roomRepository;
-
-    @Autowired
-    private CustomChecklistQuestionRepository customChecklistQuestionRepository;
 
     @Autowired
     private ChecklistOptionRepository checklistOptionRepository;
@@ -115,29 +103,6 @@ class ChecklistServiceTest extends IntegrationTestSupport {
         assertThatThrownBy(() -> checklistService.createChecklistLike(USER1, checklist.getId()))
                 .isInstanceOf(BangggoodException.class)
                 .hasMessage(ExceptionCode.LIKE_ALREADY_EXISTS.getMessage());
-    }
-
-    @DisplayName("체크리스트 질문 조회 성공")
-    @Test
-    void readChecklistQuestions() {
-        // given
-        customChecklistQuestionRepository.saveAll(CustomChecklistFixture.CUSTOM_CHECKLIST_QUESTION_DEFAULT);
-
-        // given & when
-        ChecklistQuestionsResponse checklistQuestionsResponse = checklistService.readChecklistQuestions(USER1);
-
-        // then
-        List<Integer> defaultQuestionsIds = CUSTOM_CHECKLIST_QUESTION_DEFAULT.stream()
-                .map(CustomChecklistQuestion::getQuestion)
-                .map(Question::getId)
-                .toList();
-        List<Integer> responseQuestionsIds = checklistQuestionsResponse.categories().stream()
-                .map(CategoryQuestionsResponse::questions)
-                .flatMap(Collection::stream)
-                .map(QuestionResponse::getQuestionId)
-                .toList();
-
-        assertThat(responseQuestionsIds).containsExactlyElementsOf(defaultQuestionsIds);
     }
 
     @DisplayName("작성된 체크리스트 조회 성공")
